Try to fix bad columns returned by mPulseAPI
Every now and then the mPulseAPI will return custom metric columns
named CustomMetric<index>..., instead of using the metric name.
This is a bug in the API, but we can fix it within our code by
shimming in the correct names from the repository.

This also helps when there are no results and the mPulseAPI sets
column names to `null`