Remove double quotes around negative number
mPulse has a bug where a negative number in an array of numbers will be
enclosed in double quotes, turning it into a string.

This patch removes double quotes around negative integers.  We restrict
the check to integers for now until we can find evidence that this also
happens for floating point numbers.